test: run_device_tests: add support for Zephyr builds
Add an option to run tests with Zephyr builds.

It builds the Zephyr image with zmake + adds additional configs needed
for a specific test. The list of tests is in
ec/zephyr/program/fpmcu/testcase.yaml file. The file follows a pattern
used by the twister test runner.

The the test is run with a console command - 'ztest run-all'.

The scrpit assumes the usage of ztest framework at the moment. It can be
changed in the future.

BUG=b:324574796
TEST=connect Bloonchipper via servo_micro and run
'./test/run_device_tests.py --board=bloonchipper --flasher=servo_micro
--tests abort --zephyr'

Change-Id: I3e2d4a4eb63727b8c4d936c30b2756b66a5bbf15