pdc: rts54xx: UCSI command bit alignments for v0.6.1 FW
This CL modifies the UCSI commands used in the rts54xx.c driver to
reflect the changes made in v0.6.1 realtek FW. This change is not
backwards compatible with v0.5.1 or earlier realtek FW.

This CL also modifies the rts54xx emulator to work with the updated
ucsi bit alignment changes.

BUG=b:326276531
LOW_COVERAGE_REASON=Coverage to be handled in b/325855794
TEST=verfied the following PDC commands on Brox:
1. GET_CONNECTOR_CAPABILITY
2. SET_UOR
3. SET_PDR
In addition, observed that the PDC responds much more quickly to the
set of CMDs issued during a connection event. Prior to these changes,
it was taking ~10 seconds to complete and now the time required is
~2-3 seconds.

Verified that both `./twister -s pdc/pdc.generic` and
`./twister -s pdc/pdc.rts54xx`are passing.

Change-Id: I5173251b071f463c93c7bf0bf442ed431aafe772