common/printf: Fix field width interpretation
Standard printf does not limit the field width to the precision, so
remove the check. Add exhaustive tests for (width, precision)
specification. A helper function is added to compare cros-ec formatting
results directly against standard results.

Also, this format specifier case for strings is not used in current EC
code, so there is no risk of inadvertent buffer overflows due to the
increased output size.

BUG=b:239233116
TEST='make buildall' passes
	'zmake build -a' passes
	'./twister -T zephyr/test' passes

Change-Id: I8cde162591decfd4809a2a03049ee947dcbbc66f