zephyr: fpc1050: improve sensor usage synchronization
More than one thread can communicate with the FP sensor, so there is a
need of synchronizing access to SPI interface. The synchronization
implemented in the SPI driver is not enough, because access to SPI should
be blocked until CS release, not only for a single SPI transaction.

Add additional semaphore to synchronize the access between threads.

BUG=b:248063863
TEST=Flash FPMCU on hatch, add a new finger and lock/unlock with finger
as fast as you can. Make sure there is no reboot caused by watchdog

Change-Id: I1e9772976a0d1addc52ca2d01d0ef40e919e35ee