board/hyperdebug: New flag for controlling I2C STOP condition
Slightly enhance the binary I2C protocol to allow opentitantool to
request transfers without I2C STOP condition, in order for the next
transfer to appear as a RESTART within the same "transaction".  (This is
how real I2C hosts would typically perform a logical sequence of
transfers on a possibly multi-master bus.  Producing STOP condition only
once the entire transaction is over, and the bus should be free for
other masters to use.)

BUG=none
TEST=none

Change-Id: Ifdd83837ccfcbc4de86b85a58d1babd7b38db425