test: Move sha256 sw/hw selection to board
Change https://crrev.com/c/5273877 removes the generic dep
CONFIG_SHA256, in favor of manually selecting SW or HW version of
sha256 for each. To make this work for common libraries and tests
(vboot and fpsensor) that simply depend on sha256 (sw or hw) across
different platforms (host and npcx), we must break the dependency of
the test/lib on sha256 and provide a platform dependency on the
correct sha256 implementation.

The other tests that specify the CONFIG_SHA256_SW in test_config.h
are currently host-only tests, so we can leave these CONFIG_SHA256_SW
defines as documentation, since it won't interfere with the platform's
existing CONFIG_SHA256_SW def. If these tests are added for on-board
testing, these CONFIG_SHA256_SW defs will need to be removed.

BUG=b:326982748
TEST=make hosttests
TEST=make BOARD=helipilot tests
TEST=make BOARD=bloonchipper tests

Change-Id: I1ee35603323f606646de0d0837d272a631bb3be5