cr50: commit NV when AP is off
Deep sleep cleans NV cache. Depending on timing of operations it may
happen when NV cache is not yet committed to flash.

To prevent corner cases with disorderly shutdown of AP, enable commits
on AP off event to preserve content to flash and enable sleep due to
NV state. This can help with corner case when AP is reset early in the
boot when commits are disabled.

BUG=b:262324344
TEST=make BOARD=cr50 H1_RED_BOARD=1 H1_RED_BOARD_DEEP_SLEEP=1 \
 CR50_USE_FIXED_CERT=1; tpm_test.py -p

Note that system goes into deep sleep.

Change-Id: I5c6425da57c1b64873264a60ff568f7e0d550b6c
Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>