pdc: Add `pd version` console command for TAST
Add a `pd version` console command to the PDC subsystem that returns '3'
to indicate the PDC architecture is in use. This command, despite
deviating from the `pdc` command structure, is what TAST calls to
determine the TCPM stack in use. Implementing just this one for the PDC
allows TAST to figure this out with one command rather than having to
probe multiple and wasting time on timeouts.

BUG=b:327444410
TEST=flash to Brox

Change-Id: Ie27853f97073eb11fa94b010cef0eba64ef9e709