pdc: Add console support for connector_reset
This CL adds support in the PDC API for the PDC command
CONNECTOR_RESET which can be used to trigger sending either
a hard or data reset to a port partner. In addition, a new PDC console
command `pdc conn_reset <port> [hard|data]' is added to support PD
TAST testing.

BUG=b:321266756
TEST=Verified with USB-PD analyzer that hard reset message is sent to
the port partner and the PD contract is re-established. Testing data
reset requires a USB-4 connection which is not yet functional.
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: Ia507d2aee3f4d59cdf31ab194521d7d9dabe6bd6