pdc: Add suuport for GET_VDO for disovery identity
This CL implements the RTS545 GET_VDO command and modifies both the
src/snk.attached flow to call this method when a new connection is
detected. These VDOs then are used to extract the VID/PID and Product
Type

BUG=b:319390850
TEST=Verifed that the VID, PID, and PTYPE are retrieved and matches
what is shown in the discovery identity message decoded in the total
phase trace.
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: I69ae334f34782407a78c238b4e52ea2220538232