helipilot: Refactor GPIOs to keep RO minimal
We want the amount of code and number of things defined in RO to be as
minimal as possible since RO is frozen forever. By keeping RO minimal,
we can reduce surface area for attacks and also confusion when GPIOs are
removed or renamed.

The fingerprint-related code only runs in RW, so move all
fingerprint-related GPIOs and associated code into separate files that
are only included in RW.

BUG=b:324261684, b:324292087
TEST=make buildall -j10
TEST=run_device_tests.py -b helipilot -f servo_micro

Change-Id: Id7013dbb1b08f3e50973bac07e5b93ea3c34a528