Rename ccd interface usb, since that is what it is
The iteflash ccd interface is not really ccd, it is really i2c over USB.
Rename to usb, to not freak anyone out when we use it for non-CCD
servos.

Set default block size to 128 just like i2c_pseudo (linux).

Accept usb or ccd in the flag for backwards compatibility.

BUG=b:327674515
TEST=iteflash on joxer over c2d2

Change-Id: I6f1218d7fd168aeb89035cd613cd80db6ea2a913