zephyr/emul/nx20p348x: Allow no-TCPC operation
Add a config option to not require access to a paired TCPC. The
nx20p348x PPC emulator (when used to emulate the nx20p3483) required
access to the paired TCPC. The paired TCPC instance is not known at
compile time when these are plumbed up at run time (i.e. 'is-alt' is
used).

BUG=b:247796816
TEST='./twister -T zephyr/test/rex' passes,
	CQ passes

Change-Id: I084ab10dd304e3bc1cf31061bebfaa8f843c383d