PPM: Run PPM & PDM in parallel
This CL allows the PPM and the PDM to communicate with a PDC without
interference. The concurrency is expected to be controlled by a PDC
driver. That is, if a PDC driver is not in IDLE, the caller (PDM or PPM)
receives -EBUSY.

The PPM's call to a PDC's API is synchronous. That is, the PPM will
block on each call. This approach saves a lot of code required to handle
the following problems:

* PDC does not need to call a callback on a CC event for the PPM.
  (If it's not a blocking call, the PDC would need to keep a flag (e.g.
  'run_for_ppm').
* The PPM shares the code with usermode PPM, which assumes the
  communication with a PDC is blocking. Thus, ePPM can continue to share
  the most code with uPPM.

The CL includes other changes:
* PPM saves a notification mask internally and use it to determine
  if it needs to notify the OPM on connector change events.

BUG=b:319124515
TEST=Boot Brox SKU1 with OPM.

Change-Id: Ie8fd0807824ab5e3871b8fb28eb9005532cacfff