Yaviks: Configure BC1.2 data role detection
Yaviks has a combination of the BC1.2 or without BC1.2 chip. Query
the BC1.2 chip with I2C to know if the BC1.2 chip exists or not.
Then we should disable BC1.2 data role detection if the BC1.2 chip
does not exist and we don’t need to process BC1.2 IRQ as well if
BC1.2 is not present.

BUG=none
TEST=1. On Yahiko. Attached BC1.2 charger/charger while S5 to G3.
Make sure the battery is charging.
2. On Yavkis. Attached BC1.2 charger/charger while S5 to G3.
Make sure the battery is charging.
3. On Yaviks. Attach USB devices. Make sure the device is workable.
4. On the EC console. Don’t see any BC1.2 fault.

Change-Id: I83de2f81075cf8bd2415e74a13f02f19e52c8e1c