Yaviks: Override board_usb_charger_support if BC1.2 chip is not present
Yaviks has a combination of the BC1.2 or without BC1.2 chip. Query
the BC1.2 chip with I2C to know if the BC1.2 chip exists or not.
Then override board_usb_charger_support to disable usb charger task.

BUG=none
TEST=1. On Yahiko. Attached BC1.2 charger/charger while S5 to G3.
Make sure the battery is charging.
2. On Yavkis. Attached BC1.2 charger/charger while S5 to G3.
Make sure the battery is charging.
3. On Yaviks. Attach USB devices. Make sure the device is workable.
4. On the EC console. Don’t see any BC1.2 fault.

Change-Id: Ifb313e65dfc8eeb6338758cfb5708397efc6ca25