fpsensor: Add migrate_legacy_template command
This command will be used to migrate a legacy template (that wasn't
created in a nonce session) to a modern template. The salt will be wiped
and the template will be treated as a new one, so biod can download it
again.

There are 2 differences of an old and a new format template such that we
need a special migration command:
1. We need to regenerate the match secret. The old protocol has a
   different security model and its match secret isn't well-protected.
2. The encryption materials are different. The new protocol uses nonce
   session encryption, and we need to bring the old template encrypted
   with user_id to the nonce session.

The modern biod state machine tracks the modern templates in the nonce
session. Therefore, the usage of the migration command can be thought of
as enrolling a new finger. Session needs to be established and the
template and match secret need to be fetched. But instead of interacting
with the user, the template is created by uploading a legacy template
through the migration command.

BUG=b:330294067
TEST=TBD
TEST=manual test fingerprint migration works locally

Change-Id: I22ed630211953f1cba6ae07232351b243fe3f3ab