vivaldi_kbd: switch from SYS_INIT to HOOK
Switch the vivaldi_kbd_init call from SYS_INIT to hook, this is to
ensure that the function gets called after cros_cbi_ec_init, which is
currently called from a hook as well but with higher priority.

Ideally we'd move cros_cbi_ec_init to SYS_INIT instead as that's the
Zephyr API for that, but I see that has been tried before so at this
stage I'd rather convert this to the previous behavior from
common/keyboard_vivaldi.c than risk of adding more regressions.

BUG=b:329033139
TEST=build and run on rex with some extra test print, ensure that the
     function get called after cros_cbi_ec_init

Change-Id: Iaaf8d5e74593bfc08a4485643d390a66dd082e19