zephyr: set Zephyr panic reason correctly
Set the panic reason provided by Zephyr, only if the esf structure is
not present.

Otherwise it doesn't make sense, because we check if a reason is
present, but the panic structure contains registers values provided with
the esf structure.

BUG=b:324574796
TEST=zmake build -a && ./twister
LOW_COVERAGE_REASON=Core specific line not covered, tested only on
hardware

Change-Id: Id643b04fa2f67fdd94f4391236f4c7af8785c28e