pdc: Clear PDC_SM_EVENT always when pdc_thread is woken
This CL adds k_event_clear() call to ensure that PDC_SM_EVENT doesn't
remain set in cases where a CMD is not able to be sent because a CMD
is already pending. This can happen if some host command or console
command posts a public CMD just after the PDC state machine starts
sending an internal CMD. In this sitatuion, the pdc_thread does not
sleep and then the retry_counter will no longer be a useful estimate
of a wait timeout.

BUG=b:329151707
TEST=Verfied that when testing with a usbc-->hdmi adapter, the PDC
state remains in SRC_ATTACHED as expected.
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: Icce88c50d995aef11163d3c2a8686a1e1561fd37