usb_mux: Add a mutex around enter/exit low power mode
Currently the LPM flag is set just before the USB mux and retimer are
put into low power mode. If another task is trying to initialize the USB
mux at around the same time the driver could clear the LPM flag and
enable the mux between when the LPM flag is set and the retimer enters
LPM. This results in the retimer being in LPM without the flag being set
and retimer access failures on the next connection.

This adds a mutex around LPM entry in enter_low_power_mode and LPM exit
in usb_mux_init to prevent putting the retimer in LPM after clearing the
LPM flag.

BUG=b:313363935
TEST=./twister -v -T zephyr/test, make runhosttests, manually checked
retimer errors did not occur on disconnect/connect with 100ms delay
added to enter_low_power_mode.

Change-Id: I544d836e2e241f837b5de5c369c2bcdc03730115