chip/npcx: Check the WP_IF bit before changing status register
When NPCX_INT_FLASH_SUPPORT is enabled, the status register is protected
if both WP_IF bit and SRP0 bit are enabled.

The WP_IF bit is cleared by MCU reboot and can't be cleared by software.
So if we disable HW WP, we still can't change the status register.

The conclusion is that we should check is_int_flash_protected() instead
of the WP GPIO line.

BUG=b:328066864
TEST=Compile helipilot firmware and run it on DUT with helipilot board.
     Make sure the firmware works and there are no regressions.

Change-Id: If27503b329afd8b2111f75d20ff8a33d0b733909