board/hyperdebug: Reset UARTs to default settings on reinit
Running `opentitantool transport init` is supposed to bring HyperDebug
to a known state, by first resetting everything to HyperDebug power-on
defaults using the `reinit` command, and then `opentitantool` will apply
a sequence of settings based on configuration files.

We rarely changes baud rate on HyperDebug ports, and it turns out that
these settings were not re-initialized by `reinit`.  This CL makes sure
that the four UARTs are put back to power-on default state.

BUG=none
TEST=run `make tast` after manually having enabled parity

Change-Id: I5ca38513684e0a6842878eae05d2636286197f9c