sm5803: remove VBAT overvoltage monitoring
It turns out that VBAT overvoltage monitoring on SM5803 systems will
almost never do anything useful and it's safe to not do it, so rather
than attempting to tune the thresholds to avoid spurious overvoltage
interrupts instead stop monitoring it.

For 3S systems, the current battery overvoltage limit is 13.28832V
(slightly lower than the 13.3V stated in comments because the LSb value
for VBATSNSP in 3S mode was rounded up to the nearest 100 microvolts).
Although the design voltage for a 3S battery will typically be around
11.5V, a fully-charged pack will generally go almost all the way to its
maximum charge voltage of 4.4V per cell (13.2V for the whole pack). In
the yaviks I've tested on, charge stops at VBAT=13.113V. 2S batteries
behave similarly, targeting a maximum of 4.4V per cell.

If the SM5803's analog monitoring were perfectly accurate this would be
okay, but it is not: VBAT monitoring is only specified to be ±5%
accurate, so ensuring no spurious interrupts would require a setpoint of
about 13.2/0.95=13.895V. This is nearly useless because the protection
function of the battery pack (for all battery packs I've inspected) is
specified to cut off charging at 4.45V per cell, leaving only a 50 mV
window of overvoltage that won't be handled by the battery pack. Since
uncertainty in the SM5803's measurement is about ±700 mV, assuming
uniform distribution of SM5803 measurements over the specified range
shows that only about 3.5% of battery overvoltage events would be
detected before the battery pack took action. In practice the
measurements probably have something more like a normal distribution
around the true voltage, so the actual rate of overvoltage detection
would probably be much lower than 3.5%.

Referring to the RAA489000 charger used in some systems where SM5803 may
also be used, the RAA489000 lacks any equivalent battery overvoltage
protection feature to the one present in the SM5803. Since systems with
that other charger are not known to have any issues with battery
overvoltage, this additionally indicates it's safe not to monitor for
VBAT overvoltage in the charger and instead depend on the battery pack
to do any required protection.

BUG=b:322215231
TEST=yaviks fully charges battery and stops charging when battery is
     full, as expected

Change-Id: Iaf4e2d036e0545b22ff9bc8cb2615c707f7d8fed