helipilot: Change rollback region size to 64KB
Currently, rollback region size on helipilot is 128KB. It turned out
that the value was copied from bloonchipper config.

In fact, the rollback region only keeps small structure, but the
requirement is that rollback regions must be placed in different erase
units, to prevent from situation that we will lose part of encryption
key when e.g. MCU was turned off while updating region.

Erase unit is set to 64KB, so set rollback region size to that.

BUG=b:329412639
TEST=Compile helipilot firmware and run it on quincy board.
     Boot FPMCU to RO using 'reboot ro'.
     Add rollback entropy using `rollbackaddent` and confirm that second
     region was updated:
     rollback 0: 00000000 00000000 0b112233 [00..00]
     rollback 1: 00000001 00000000 0b112233 [31..4a] *

     Add rollback region again. Confirm that the first region was
     updated and the second one was left intact.
     rollback 0: 00000002 00000000 0b112233 [d9..f3] *
     rollback 1: 00000001 00000000 0b112233 [31..4a]
TEST=Run TAST tests and make sure the tests pass.
TEST=./test/run_device_tests.py --board=helipilot \
     --flasher=servo_micro --with_private=no --tests rollback.*

Change-Id: I6195c7453f0f5599385054f18286177f69eb9306