geralt: Don't disable lid detection when base attached
Considering the firmware.FWScreenCloseLid.broken, after entering
RECOVERY screen and calling lid close, ec will issue
EC_REBOOT_COLD_AP_OFF. If we init lid detection disabled, and then
the lid open issued at this time, it won't be able to boot the DUT,
resulting in the test failed.

For the CONFIG_GERALT_LID_DETECTION_SELECTED, it should be worked for
base detached. So we should not disable lid detection when base
attached.

BUG=b:328387518
TEST=tast run tast.firmware.FWScreenCloseLid.broken

Change-Id: I76ade7f019ec67f095379f7e47a9bedc498c9cb7