zephyr/emul/rts54xx: Fix read result with offset handling
The pdc driver GET_VBUS_VOLTAGE command is implemented using a partial
read of the PDC GET_RTK_STATUS result message. Result messages always
start with a byte count followed by the indicated number of bytes.

This means we need to treat GET_RTK_STATUS with a non-zero offset as a
special case since we need to respond with the byte count from byte 0 of
the full response and additional bytes at the requested offset.

for test_get_bus_voltage, rts5453p_emul_read_byte now records:

  read_byte reg=0x80, bytes=0, offset=18, val=0x2
  read_byte reg=0x80, bytes=1, offset=18, val=0x64
  read_byte reg=0x80, bytes=2, offset=18, val=0x0

which is the expected result message.

BUG=b:330414698
TEST='./twister -T zephyr/test/pdc' passes

Change-Id: Icfd62c85663832e50ac4d2f0462ebbc3aade3601