pdc: Setting data IRQ to level trigger
Setting data IRQ to level trigger. The interrupt
will assert as long as the the IRQ line is low.

This mitigates the issue where the PDC doesn't
recognize a clear IRQ command and doesn't reassert
the IRQ.

BUG=b:305014156,b:325624686
TEST=DP-ALT mode works as expected. Verified with scope
that interrupt line is asserted and cleared.
LOW_COVERAGE_REASON=Coverage to be handled in b/325855794

Change-Id: Ic34c105cd5e3574ee9f4ce833e18c36e9aa18dde