altmode: add logging for all mux changes
The value read from the DATA_STATUS register is helpful for debugging,
so log the value on any changes. Data is displayed MSB first.

The "altmode read" output is also change to match the format of the
logging.

Output:
I: P1 DATA_STATUS: 00 00 00 85 51

BUG=none
TEST=boot EC image on brox
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: Icf816d13323e3d41b75b9efc5628c68ce3877e8c