pdc: Add named-pdc-port node
The named-usbc-port node was originally introduced for TCPMv2 and has
since been expanding for use in PDC-based devices. This, however, means
there any many properties that are mutually exclusive and it isn't clear
which are needed for the PDC vs TCPMv2 vs are common. To help maintain a
clearer separation between these two PD solutions, introduce a
named-pdc-port node that behaves similarly but is PDC-specific and
remove those properties from named-usb-node.

BUG=b:330401034
TEST=zmake build brox brya --clobber

Change-Id: I3266a46a81fca9cbfb8dd85045151ef7a3a83861