pdc: Modify data structs used to get SRC/SNK caps
This CL modifies the data structs used for retrieving SRC/SNK caps from
the PDC. This provides the ability to read SRC or SNK caps from either
the port partner or the PDC (LPM).

BUG=b:319643480
TEST=Verfied on Brox that SRC CAPs can be read from both the port
partner and the PDC's LPM. Also verifed that SNK CAPs are from the
port partner as expected.
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: I880accb8a0cf10bfa2b0796e93ca220b471890bd