pdc: Fix issue with RTS54 SET_TPC_RP cmd format
This CL modifies the SET_TPC_RP driver function to set both the TPC Rp
and PD Rp fields. The documentation doesn't state this, but
setting the TPC Rp field only results in an error from the TPC.

BUG=:
TEST=Verified that sending SET_TPC_RP no longers results in an error
ping status result. In addition, verified with total_phase that CC
voltage level does change as expected when the Rp value is changed.

Change-Id: I12c4a2623d4a574459fb3041518ad2326004cfbd