cr50: stop building private* boards
We're working on unnesting the private* repos. This could temporariliy
break building the private* boards while everything is landing.

The CQ depends on make buildall. Stop building the private boards in
make buildall until we're sure everything works.

This is only temporary. Revert this CL after the symlinks to the
private* repos are up.

BUG=b:323962993
TEST=make buildall -j ; ls build/ doesn't show any private boards

Change-Id: Id7d064d44b08a87c5bd75a8c76787c2e35bb9e1a