keyboard: Introduce KEYBOARD_COLS
Currently, KEYBOARD_COLS_MAX is used to define the default scancode
table and the arrays holding the scanner's states. These sizes are not
necessarily equal if FW supports two keyboards of different sizes.

This CL introduces KEYBOARD_COLS to decouple the two sizes. It indicates
the column size of the default matrix while KEYBOARD_COLS_MAX continues
to indicate the column size of the largest scancode table (as inferred
by the name). It's used to allocate the arrays holding the scanner
states.

The runtime column size is tracked by keyboard_cols. The scanner uses it
as a loop terminater.

BUG=b:299186434
TEST=Xol

Change-Id: Ibca46ce4df87d2cbf1a5c086dbce18a97080107a