keyboard: Introduce KEYBOARD_COLS
Currently, KEYBOARD_COLS_MAX is used to define the default scancode
table and the arrays holding the scanner's states. These are not
necessarily true if FW supports two keyboards of different sizes.

This CL introduces KEYBOARD_COLS, which indicates the column size of the
default matrix. KEYBOARD_COLS_MAX will be now used only as the column
size of the largest matrix (as the name indicates). It's used to
allocate the arrays holding the scanner states.

keyboard_cols holds a runtime column size. The scanner uses it as a loop
terminater.

BUG=b:299186434
TEST=Xol

Change-Id: Ibca46ce4df87d2cbf1a5c086dbce18a97080107a