fpmcu: Set C standard to C11
If the C standard (CSTD variable) is not set, Zephyr uses C99 by
default.

If C99 standard is set, BUILD_ASSERT() macro is a no-op on Clang,
because there is a check if __STDC_VERSION__ >= 201100. Obviously,
this check is not satisfied when C99 standard is used because compiler
pretends to be C99 compiler.

Starting from C11, compilers provide _Static_assert() builtin used to
implement BUILD_ASSERT().

BUG=b:331380812
TEST=Add BUILD_ASSERT(false) to some file compiled when building
     bloonchipper and make sure compilation fails

Change-Id: I19059646ff2784cc2491c154ddf260bc193b035b