test/aes: Use public BoringSSL APIs
Note that this change also needs different test vectors since the public
API performs checks on the input, such as the length of the key.

The new test vectors comes from
https://boringssl.googlesource.com/boringssl/+/f94f3ed3965ea033001fb9ae006084eee408b861/crypto/cipher_extra/test/aes_128_gcm_tests.txt

Prior to this change, the AES-GCM encrypt and AES-GCM decrypt benchmarks
in this test took an average of ~4060 us per iteration on helipilot.
After this change, they now take ~2270 us per iteration.

BUG=b:279950931
TEST=make run-aes
TEST=./test/run_device_tests.py --board helipilot -t aes

Change-Id: I65dd0e63681c4baf39759be58aa89d91ff55f7ee