chip/stm32: Fix for "deadlock" in UART forwarding
uServo, C2D2 or HyperDebug can get into a state in which the host
computer does not receive any serial characters via USB, despite the DUT
producing output on its UART.

The above happens if the character queue is filled and overflow BEFORE a
USB connection is established.  As the queue is already full, further
UART activity from the DUT does not result in the queue consumer being
notified.  On the consumer end, the queue is inspected only in the
function `usb_stream_deferred()`, which is called after every USB
transfer.  However, when the USB connection was established
`usb_stream_event()` would have set the TX endpoint to NAK (meaning no
data available for the host to receive), so there will be no USB
activity, and hence `usb_stream_deferred()` will never be called.  (At
least, not until a character is sent "the other way", by e.g. typing
in a CCD prompt.)

This CL modifies `usb_stream_event()` such that on USB connection
establishment/reset, it will inspect the character queue, and set the TX
endpoint to either VALID or NAK, depending on whether characters are
available in the queue.

BRANCH=servo
BUG=https://github.com/lowRISC/opentitan/issues/19564
TEST=observe HyperDebug able to perform serial flash of Ti50
TEST=make tast TAST_EXPR=gscdevboard.GSCTPM.*  (in ti50/common)
Change-Id: Ia1e44ff9e85474768a38b4ced1200fc7c5d05875

Change-Id: I5eb9c425afe6c1b7738253e9878c4e59299e0354