ppc/syv682x: Add use of vbus to trigger FRS
In SYV682, there are two ways to determine FRS:
a. When CC_FRS = 0, the judgment of FRS is CC low.
b. When CC_FRS = 1, the judgment of FRS is VBUS.
Currently we use CC_FRS = 0 by default and do not use CC_FRS = 1.
In order to avoid overlapping trigger conditions for FRS, a way
for vbus to determine FRS is added. Enable FRS via
board_pd_set_frs_enable.

BUG=None
TEST=zmake testall

Change-Id: I629488f542fe6f36ad602f5b9f188f602cf2c474