PPM: Add enable switch in CBI common control
This CL adds a flag to the common control tag (i.e.
CBI_TAG_COMMON_CONTROL), which disables the UCSI stack if the flag is
set. When the UCSI stack is disabled, the CrOS PD stack is expected
to become active.

$ ectool cbi get 11
EC result 3 (INVALID_PARAM)
Error code: -1003

$ ectool cbi set 11 1 4

$ ectool cbi get 11
As uint: 1 (0x1)
As binary: 01 00 00 00

BUG=b:319124515
TEST=Brox SKU1. See above.

Change-Id: I524ce09f99bc68ec8dab70ae6820b8641b895dce