PPM: Rename ucsi_disabled to ucsi_enabled
A UCSI switch was added by crrev/c/5421069. This CL reverses the logic.
That is, if ucsi_enabled is set, the UCSI stack will be enabled. If it's
not set or COMMON_CONTROL doesn't exist, the UCSI stack will be
disabled.

When the transition completes, we'll remove the code reading this switch
(from AP FW and EC FW). After that, all devices will have the UCSI stack
enabled regardless of the COMMON_CONTROL value.

BUG=b:319124515
TEST=Brox SKU1

Change-Id: I05c4bf4bb3c3a86c05afbd9b4c4600cf012ee3c7