pdc: Don't report PDC_SEND_CMD_START / PDC_SEND_CMD_WAIT externally
The `pdc status` console command should not indicate the PDC SM's
transitional PDC_SEND_CMD_START and _WAIT states as they are not actual
PD states. If we are in one of these states, report the return-to state
instead. This reduces the risk of TAST tests getting confused.

BUG=b:332716850
TEST=No SendCmdStart/Wait seen while plugging in Type-C cables
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: Ib1a99474961968d7146e6b538783aeaa5a093af4