test/aes: Use hex-encoded test vectors
This makes it easier to use the test vectors directly from BoringSSL
rather than having to convert them:
https://boringssl.googlesource.com/boringssl/+/f94f3ed3965ea033001fb9ae006084eee408b861/crypto/fipsmodule/modes/gcm_tests.txt

To make sure that the hex-encoded test vectors are exactly the same as
the original test vectors, this change compares the decoded hex test
vectors to the original vectors. A followup commit will remove the
original test vectors.

BUG=b:279950931
TEST=make run-aes

Change-Id: Ie4b1a8c9089079a765c68821ad6fabe010d10970