brox-ish-ec: config tablet mode switch
1. Config TABLET_MODE_SWITCH so EC could send mkbp switch event to AP
2. Wait for proto 2 to enable tablet mode interrupt node
This node handles interrupt from ISH, but signal is not wired to EC
on proto 1; hardware change is applied to proto 2. So remove this
node for now and enable it when proto 2 is ready.

BUG=b:319165213 b:332805291
TEST=After system boots up, on EC console,
"tabletmode" shows clamshell;
"tabletmode on", screen rotation works, and keyboard doesn't accept
input; "tabletmode off", screen won't rotate, keyboard works.

Change-Id: Ib3f38a0ca1c582d2c07742967c81695629638384