zephyr/test/pdc: Add PDC message tracing
This enables the PDC message tracing (CONFIG_USBC_PDC_TRACE_MSG) feature
for PDC tests. PDC messages are reported in the test logs as well as
captured in a PCAP file for further analysis. The new PCAP file is
called rts.pcap and located in each test's output directory next
to build.log and handler.log.

The TEST_PDC_PROTOCOL_TRACING Kconfig option is introduced for enabling
this feature in tests.

BUG=b:336660528
TEST='./twister -T zephyr/test/pdc' passes
	new "rts.pcap" file present in each test's output dir.

Change-Id: If7c08a6bc0dccdd30a069c3d2091d67d23809a9e