UCSI/PPM: Fix UCSI command logging
Currently, the cmd_val variable in rts5453_smbus_command is used to
determine which logline gets printed out when the PPM is sending an
smbus command. But, the cmd_val is compared directly to SC_UCSI_COMMANDS
which is false for UCSI commands. This updates the condition to compare
cmd with the SC_UCSI_COMMANDS to properly print out the UCSI commands
sent by the PPM.

BUG=None
TEST=make um_ppm, checked ucsi commands were decoded in verbose um_ppm
logs.

Change-Id: Ia072eb95dc8d3455911fc0f906626af319aa88bc