pd_controller: enable debug commands for PDC-PMC
To troubleshoot PDC-PMC issues, this patch enables the debug
commands 'altmode' and 'ectool usbpdmuxinfo'.
This is implemented by sending VDM command GET_PCH_DATA_STATUS
to PDC and retrieving the required information.
The information received is populated in the debug commands.

BUG=b:330475942
TEST=zmake build brox;
LOW_COVERAGE_REASON=Coverage to be handled in b/337721162

Change-Id: I19b2657e5c90f9cf0006553a3c438ee371ce5173