ectool: Add PDC trace command
Add support PDC message tracing. Messages are retrieved from the EC and
forwarded to either a network destination or PCAP file.

CONFIG_USBC_PDC_TRACE_MSG must be enabled on the EC for this to work.

The "ectool pdtrace ..." command is enabled with these options:

  -p <PORT>
    collect on USB-C port <PORT>
  -d <HOST>
    send messages to host <HOST>
  -s
    write message to stdout
  -v
    send messages to the running crostini instance (on-device only)
  -w <FILE>
    write messages to file <FILE> in PCAP format

The wireshark plugin for analyzing PDC trace messages is introduced in
chromium:5536339.

BUG=b:336660528
TEST=exercised on brox

Change-Id: Idebf906092ea85343b00806ad412ad3009955a6e