drivers/usbc/pdc: PDC message trace implementation
Add feature to enable the AP to trace communication between the EC and
the PDC (power delivery controller). Hooks in the PDC receive and
transmit paths are utilized to capture messages in a FIFO that the AP
(e.g. ectool) can interrogate using host commands.

Tracing is disabled by default and only enabled when requested by the AP
via host commands.

link: go/cros-on-device-pdc-wireshark

BUG=b:336660528
TEST=tested using ectool and EC console

Change-Id: I0e66438773ff4a9335363aa951b259b40705640d