tcpmv2: Don't call tcpm_debug_detach during PRS
The tcpm_debug_detach is currently implemented in only one TCPC driver -
ra489000. It applies CC_OPEN in order for the TCPC state machine to
detect the debug accessory being detached.
Now the problem here is that it also seems to break PRS with DTS on.
Fix that by only calling tcpm_debug_detach if we're not in the middle of
doing PRS.

BUG=b:312370715
TEST=firmware.ECPDPowerSwap.normal passes on pujjoteen
LOW_COVERAGE_REASON=b:153071799

Change-Id: I5fc79a63d7863f6a989c00093fbcbc8aa7e5a8fb