test/run_device_tests.py: Expand helipilot idle power range
This is not the sleep range that we truly care about.
This expands the "correct" idle power range from between 32mW and 38mW
to anywhere between 28mW and 42mW.

Honestly, I don't know if we actually found the center of the power
consumption distribution, so adding a larger offset is the insurance
policy. As another data point, my quincy board MCU hovers around 30mW.

BUG=b:339491623, b:332782430
TEST=./test/run_device_tests.py --board=helipilot \
       --flasher=servo_micro --with_private=no \
       --test power_utilization_idle

Change-Id: I4299a233238fb5050c89c29332478c0cd4ebfb28