board/hyperdebug: Clear UART buffers on reinit
When Hyperdebug receives the command `reinit`, the caller expects the
device to restore all settings as they were on powerup.  Generally the
best way to achieve that would have been to perform an actual reset, but
we want to maintain the USB connection, do instead various modules
register a "reinit" hook, to as best they can restore the settings.

Among the existing hooks are one to reset the four UARTs to 115.2 kbaud.
This CL further adds code to discard any characters in inbound or
outbound queue or FIFO.  (It would not make sense to continue to output
pre-queued characters at the "reset" baud rate instead of the setting
that applied when they were enqueued, also, since the caller asks for a
re-initialization, they are not interested in previously received and
queued characters.)

BUG=none
TEST=none

Change-Id: Iaaec354c9d555bac24764d15495b8af150c1d1d5