board/hyperdebug: Bring CMSIS-DAP to initial state on reinit
The binary protocol extensions used by HyperDebug on the CMSIS-DAP
interface sometimes use long messages, and the CMSIS_DAP task may block
waiting for more data from the USB host, or block waiting for the USB
host to receive queued data, making room for more.

Occationally, an opentitantool process may be terminated without
completing such a request, leaving HyperDebug in a state of not being
able to properly respond to further CMSIS-DAP requests.

This CL adds code such that the `reinit` command will signal the
CMSIS_DAP task to abort any reading/writing, and "unwind", that is,
return to the main event loop.

BUG=none
TEST=manually sent incomplete CMSIS request, then `make tast`

Change-Id: Iab488d0f715b62497aeba43bde34c3ff314bcf75