board/c2d2: Avoid glitch on UART TX when releasing reset
OpenTitan GSC enters "rescue" bootloader mode by its UART RX line being
held low while coming out of reset.  We want to be able to do this on
Chromebooks with either C2D2 or uServo connection.  Recently, a feature
was added to the EC codebase to generate "break condition" on UART pins,
but it turns out that such a break on C2D2 is briefly interrupted as
voltage level converters are disabled in the milliseconds after
releasing reset, in order to allow detecting the voltage of GSC/EC.

This is a problem, since the OpenTitan chip would be sampling the UART
line shortly after releasing reset.

This CL adds a workaround in that if a UART break condition is
active (C2D2 driving its UART TX low indefinitely) while C2D2 is
instructed to release GSC reset, then it will skip voltage level
detection, and continue using the voltage levels detected before
applying reset.

BUG=b:319123668
TEST=opentitantool --interface c2d2 rescue firmware ...

Change-Id: I1b343d0b56f911be7318c333d0da38cca70ea510