pdc: Add delay substate for type-c only connections
This CL adds a new substate for both the PDC_SRC_TYPEC_ONLY and
PDC_SNK_TYPEC_ONLY states to add delay to allow for a PD contract to
be established and the connection type be updated by the PDC.

The PDC will always report the type-c attach, and then later report a
USB-PD attach. The delay allows the PDC subsys to fully differentiate
a type-c only connection for policies such as the port's source
current limit.

The delay is inserted after the substates where the source current is
enabled or the sink path is enabled/disabled so there is no additional
delay being added for those steps.

BUG=b:319643480
TEST=Test on Brox with both type-c only and USB-PD devices and
validated that USB-PD devices don't reach the final substate, but
type-c only devices do after the specified delay has been reached.
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: Ie36695c9285fe6bdd1533e32f98d5d6a77ec91ef