kelpie: fix key-ro section offset
CL:5380581 removed PSTATE region and increased RO by 4kB.
This causes RO update failure when updating from an EC version before
CL:5380581 to a newer version, because the updater tries to write
key-ro data into an offset that is outside of old RO region.

To workaround this, move key-ro to its original location and leave
the pstate region empty.

BUG=b:319924645
TEST=able to update firmware from old preflash FW
     (kelpie-0.0.0-8290868+ 2024-01-08) to ToT.

Change-Id: I94f5e8c10f4f811fa13f22385bedda90f0b41e02