brox: remove EC_CMD_GET_PD_PORT_CAPS support for PMC-PD
host command EC_CMD_GET_PD_PORT_CAPS support removed
for PMC-PD as this is used by the the AP firmware to fetch
the USBC port capabilities to add 'cros-ec-typec' device
entry in the ACPI table. 'cros-ec-typec' module is only
essential for legacy implementation. This module querries
EC for USBC related information and configures PMC accordingly.
As PMC directly communicates with PDC, 'cros-ec-typec'
support is not required.

BUG=b:327622474
TEST=zmake build brox;make build all;
LOW_COVERAGE_REASON=Coverage to be handled in b/337721162

Change-Id: Ia98fd43b32697f99dbff0bec601c876265738284